// src/services/api/generated/models/test-run-status-enum.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Test Management API
 * API for managing boards, tests, and execution.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * * `PENDING` - Pending * `RUNNING` - Running * `PAUSED` - Paused * `COMPLETED` - Completed * `FAILED` - Failed * `KILLED` - Killed
 */

export const TestRunStatusEnum = {
    PENDING: 'PENDING',
    RUNNING: 'RUNNING',
    PAUSED: 'PAUSED',
    COMPLETED: 'COMPLETED',
    FAILED: 'FAILED',
    KILLED: 'KILLED'
} as const;

export type TestRunStatusEnum = typeof TestRunStatusEnum[keyof typeof TestRunStatusEnum];